+++ CHANGES.md
@@ -40,6 +40,15 @@ To be released.
     `string?` to `IValue?`.  [[#3111]]
  -  Changed the return type for `IActionTypeLoader.Load()` from
     `IDictionary<string, Type>` to `IDictionary<IValue, Type>`.  [[#3111]]
+ -  Changed return types and parameter types of serveral methods from
+    `IReadOnlyList<ActionEvaluation>` to `IReadOnlyList<IActionEvaluation>`.
+    [[#3089]]
+     -  `BlockChain<T>.DetermineGenesisStateRootHash()`
+        method's `evaluations` parameter type.
+     -  `BlockChain<T>.DetermineBlockStateRootHash()`
+        method's `evaluations` parameter type.
+     -  `BlockChain<T>.EvaluateBlock()` method's return type.
+     -  `BlockChain<T>.EvaluateGenesis()` method's return type.
 
 ### Backward-incompatible network protocol changes
 
@@ -49,6 +58,7 @@ To be released.
 
  -  Added `IActonEvaluator` interface.  [[#3082]]
  -  Added `ActionTypeAttribute(int)` constructor.  [[#3111]]
+ -  Added `IActionEvaluation` interface.  [[#3089]]
 
 ### Behavioral changes
 
@@ -72,6 +82,7 @@ To be released.
 [#3082]: https://github.com/planetarium/libplanet/pull/3082
 [#3083]: https://github.com/planetarium/libplanet/pull/3083
 [#3087]: https://github.com/planetarium/libplanet/pull/3087
+[#3089]: https://github.com/planetarium/libplanet/pull/3089
 [#3092]: https://github.com/planetarium/libplanet/pull/3092
 [#3098]: https://github.com/planetarium/libplanet/pull/3098
 [#3106]: https://github.com/planetarium/libplanet/pull/3106